+++ utility/distance.h
@@ -2,6 +2,6 @@
 #include "position.h"
 #include <algorithm>
 
-double ManhattanDistance(Position node, Position goal);
-double DiagonalDistance(Position node, Position goal);
-double EuclideanDistance(Position node, Position goal);
+float ManhattanDistance(Position node, Position goal);
+float DiagonalDistance(Position node, Position goal);
+float EuclideanDistance(Position node, Position goal);
